main と App をできるだけ汚さないようにする
